agent-core/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
httpx==0.26.0
websockets==12.0
python-dotenv==1.0.1
openai==1.12.0
numpy==1.26.4
scikit-learn==1.4.0
tiktoken==0.6.0
aiofiles==23.2.1
python-multipart==0.0.9
